+++ src/Gui/Dot.cpp
@@ -97,8 +97,8 @@ namespace BlendInt {
 			}
 
 			case FormRoundRadius: {
-				const int* radius_p =
-								static_cast<const int*>(request.data());
+				const float* radius_p =
+								static_cast<const float*>(request.data());
 				VertexTool tool;
 				tool.Setup(size(), DefaultBorderWidth(), round_type(), *radius_p);
 				tool.UpdateInnerBuffer(m_inner_buffer.get());
